Fix buffer length complaints; support posh-git
I was getting complaints about setting the buffer size on init, this change fixes it. 

I also changed the prompt regex.  Making the posh-git custom prompt play nicely with shell mode was sort of annoying because it mysteriously splits off the "PS " from the start of the prompt.  I also think it's better form to use  [^>]*>  instead of [^%$>]*> since we're mandating the prompt terminates with an angle bracket either way.